% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_job.R
\name{create_job}
\alias{create_job}
\title{create_job}
\usage{
create_job(path, subname, jobname, projectletter, qletter, qlength, R = F)
}
\arguments{
\item{path}{path to directory where you want the submission script to be created}

\item{subname}{name of the script you would like to run, minus the extension R of sh}

\item{jobname}{name of the job submitted to the cluster}

\item{projectletter}{letter of the project you would like to use}

\item{qletter}{letter of the queue you would like to use}

\item{qlength}{length of the job you would like to submit e.g. short}

\item{R}{is the job an R script? If not, defaults to bash script}
}
\value{
creates a file to run through the cluster with qsub
}
\description{
Creates a simple submission for the BMRC cluster
}
\details{
This is an old function, barely used anymore
}
\author{
Jamie Inshaw
}
